rag_pipeline/requirements.txt
requests
beautifulsoup4
supabase
google-cloud-aiplatform
pytesseract
Pillow
opencv-python
PyMuPDF
